requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pillow==10.3.0
aiofiles==23.2.1
tqdm==4.66.4
matplotlib==3.9.0
seaborn==0.13.2
numpy==1.26.4
pandas==2.2.2
opencv-python-headless==4.10.0.82
requests==2.32.2
dill==0.3.8
gitpython==3.1.43
setuptools==69.5.1